requirements.txt
pandas>=1.5.0
openpyxl>=3.0.0
xlrd>=2.0.0
python-dotenv>=0.19.0
email-validator>=1.1.3 